+++ frontend/wailsjs/go/main/App.d.ts
@@ -26,6 +26,8 @@ export function OpenRepositoryOnGitHub(arg1:main.Repo):Promise<void>;
 
 export function OpenURL(arg1:string):Promise<void>;
 
+export function PullLatest(arg1:string):Promise<string>;
+
 export function PullRequests(arg1:string):Promise<Array<main.PullRequest>>;
 
 export function ReadFile(arg1:string,arg2:string,arg3:string):Promise<main.Document>;